fixup! Fix always return False in extract_jar

+++ helper/download.py
@@ -119,5 +119,7 @@ def download_framework(
         target_file = os.path.join(dest_dir, target_file)
         dest_dir = os.path.join(dest_dir, filename)
         file_path = os.path.join(dest_dir, f"{filename}.{app_type}")
-        os.rmdir(os.path.dirname(file_path))
-        return extract_jar(file_path, target_file)
+        if extract_jar(file_path, target_file):
+            os.rmdir(os.path.dirname(file_path))
+            return True
+        return False
